DAG design ideas
